+++ packages/core/src/render3/view_engine_compatibility.ts
@@ -6,48 +6,14 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {ChangeDetectorRef as ViewEngine_ChangeDetectorRef} from '../change_detection/change_detector_ref';
 import {Renderer2} from '../render/api';
-import {TNode, TNodeType} from './interfaces/node';
 import {isProceduralRenderer} from './interfaces/renderer';
-import {isComponentHost, isLView} from './interfaces/type_checks';
-import {DECLARATION_COMPONENT_VIEW, LView, RENDERER} from './interfaces/view';
+import {isLView} from './interfaces/type_checks';
+import {LView, RENDERER} from './interfaces/view';
 import {getCurrentTNode, getLView} from './state';
 import {getComponentLViewByIndex} from './util/view_utils';
-import {ViewRef} from './view_ref';
 
 
-
-/** Returns a ChangeDetectorRef (a.k.a. a ViewRef) */
-export function injectChangeDetectorRef(isPipe = false): ViewEngine_ChangeDetectorRef {
-  return createViewRef(getCurrentTNode()!, getLView(), isPipe);
-}
-
-/**
- * Creates a ViewRef and stores it on the injector as ChangeDetectorRef (public alias).
- *
- * @param tNode The node that is requesting a ChangeDetectorRef
- * @param lView The view to which the node belongs
- * @param isPipe Whether the view is being injected into a pipe.
- * @returns The ChangeDetectorRef to use
- */
-function createViewRef(tNode: TNode, lView: LView, isPipe: boolean): ViewEngine_ChangeDetectorRef {
-  // `isComponentView` will be true for Component and Directives (but not for Pipes).
-  // See https://github.com/angular/angular/pull/33072 for proper fix
-  const isComponentView = !isPipe && isComponentHost(tNode);
-  if (isComponentView) {
-    // The LView represents the location where the component is declared.
-    // Instead we want the LView for the component View and so we need to look it up.
-    const componentView = getComponentLViewByIndex(tNode.index, lView);  // look down
-    return new ViewRef(componentView, componentView);
-  } else if (tNode.type & (TNodeType.AnyRNode | TNodeType.AnyContainer | TNodeType.Icu)) {
-    // The LView represents the location where the injection is requested from.
-    // We need to locate the containing LView (in case where the `lView` is an embedded view)
-    const hostComponentView = lView[DECLARATION_COMPONENT_VIEW];  // look up
-    return new ViewRef(hostComponentView, lView);
-  }
-  return null!;
-}
 
 /** Returns a Renderer2 (or throws when application was bootstrapped with Renderer3) */
 function getOrCreateRenderer2(view: LView): Renderer2 {
