+++ packages/core/src/render3/view_engine_compatibility_prebound.ts
@@ -7,13 +7,12 @@
  */
 
 
-import {ChangeDetectorRef} from '../change_detection/change_detector_ref';
+import {ChangeDetectorRef, injectChangeDetectorRef} from '../change_detection/change_detector_ref';
 import {InjectFlags} from '../di/interface/injector';
 import {createTemplateRef, TemplateRef} from '../linker/template_ref';
 import {throwProviderNotFoundError} from './errors';
 import {TNode} from './interfaces/node';
 import {LView} from './interfaces/view';
-import {injectChangeDetectorRef} from './view_engine_compatibility';
 
 
 /**
